--- kucherx/domain/queue_quit_object.py
@@ -1,2 +1,0 @@
-class QueueQuitObject:
-    pass
