+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name            = KucherX
-version         = file: kucherx/VERSION
+name            = Yukon
+version         = file: yukon/VERSION
 author          = Zubax Robotics
 url             = https://zubax.com
 description     = A simple user-friendly GUI application designed for configuration, diagnostics, and maintenance of Cyphal-enabled Zubax hardware.
