+++ setup.py
@@ -9,7 +9,7 @@ import setuptools
 import distutils.command.build_py
 from pathlib import Path
 
-PACKAGE_NAME = "kucherx"
+PACKAGE_NAME = "yukon"
 DSDL_SOURCE_ROOT = Path(__file__).resolve().parent / PACKAGE_NAME / "dsdl_src"
