+++ yukon/domain/_expand_subjects.py
