+++ yukon/domain/avatar.py
@@ -7,12 +7,12 @@ from typing import Optional, Any, Callable
 import pycyphal
 from pycyphal.transport import ServiceDataSpecifier, Timestamp, AlienTransfer, MessageDataSpecifier
 
-from kucherx.domain.note_state import NodeState
-from kucherx.domain.port_set import PortSet
-from kucherx.domain._expand_subjects import expand_subjects, expand_mask
-from kucherx.domain.iface import Iface
+from yukon.domain.note_state import NodeState
+from yukon.domain.port_set import PortSet
+from yukon.domain._expand_subjects import expand_subjects, expand_mask
+from yukon.domain.iface import Iface
 import uavcan
-from kucherx.services.value_utils import _simplify_value, explode_value
+from yukon.services.value_utils import _simplify_value, explode_value
 
 logger = logging.getLogger()
 logger.setLevel("ERROR")
