+++ yukon/domain/note_state.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 import uavcan
 from uavcan import node
-from kucherx.domain.port_set import PortSet
+from yukon.domain.port_set import PortSet
 
 
 @dataclasses.dataclass(frozen=True)
