+++ yukon/html/monitor/monitor.js
