+++ yukon/html/monitor/style.css
