+++ yukon/html/monitor/tabs.js
