+++ yukon/services/cyphal_worker.py
@@ -1,15 +1,14 @@
 import asyncio
 import logging
 import traceback
-from multiprocessing.managers import ValueProxy
 
 from pycyphal.application import make_node, NodeInfo, make_transport
 
 import uavcan
-from kucherx.domain.attach_transport_request import AttachTransportRequest
-from kucherx.domain.attach_transport_response import AttachTransportResponse
-from kucherx.domain.god_state import GodState
-from kucherx.services.make_tracers_trackers import make_tracers_trackers
+from yukon.domain.attach_transport_request import AttachTransportRequest
+from yukon.domain.attach_transport_response import AttachTransportResponse
+from yukon.domain.god_state import GodState
+from yukon.services.make_tracers_trackers import make_tracers_trackers
 
 logger = logging.getLogger(__name__)
 logger.setLevel("NOTSET")
